+++ CHANGELOG.md
@@ -7,6 +7,52 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-10-16
+
+### Added
+
+- New `reload` method to refresh model state from the database.
+
+## [0.7.2] - 2024-06-17
+
+### Fixed
+
+- Ensure TTL policies are created by `async_set_up_composite_indexes_and_ttl_policies`
+  and `set_up_composite_indexes_and_ttl_policies` also when passing in the models as a
+  generator, for example when using `get_all_subclasses()`.
+
+## [0.7.1] - 2024-06-14
+
+### Fixed
+
+- Don't raise an `AttributeError` when setting up indexes for models if there's a model
+  without any indexes (i.e. a model does not at all define the `__composite_indexes__`).
+- Fix bug that the configured prefix was not used when creating indexes. Please note
+  that if you have been using indexes and a collection name prefix, the indexes created
+  before this fix will be for the wrong collection names (i.e. missing the prefix)! Thus
+  please go through your indexes and remove the accidentally created ones after updating
+  to this version.
+
+## [0.7.0] - 2024-03-27
+
+### Added
+
+- Support for composite indexes via `__composite_indexes__` property in model classes.
+
+## [0.6.0] - 2024-02-26
+
+### Added
+
+- Add support for order_by, limit and offset in find queries
+
+## [0.5.1] - 2024-02-12
+
+### Changed
+
+- Fix pytest warnings in console output
+- Update CI pipeline to use trusted PyPI publisher instead of a token
+- Add make-changelog invoke command
+
 ## [0.5.0] - 2023-10-09
 
 ### Changed
@@ -198,7 +244,13 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Update README.md
 - Update .gitignore
 
-[unreleased]: https://github.com/ioxiocom/firedantic/compare/0.5.0...HEAD
+[unreleased]: https://github.com/ioxiocom/firedantic/compare/0.8.0...HEAD
+[0.8.0]: https://github.com/ioxiocom/firedantic/compare/0.7.2...0.8.0
+[0.7.2]: https://github.com/ioxiocom/firedantic/compare/0.7.1...0.7.2
+[0.7.1]: https://github.com/ioxiocom/firedantic/compare/0.7.0...0.7.1
+[0.7.0]: https://github.com/ioxiocom/firedantic/compare/0.6.0...0.7.0
+[0.6.0]: https://github.com/ioxiocom/firedantic/compare/0.5.1...0.6.0
+[0.5.1]: https://github.com/ioxiocom/firedantic/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/ioxiocom/firedantic/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/ioxiocom/firedantic/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/ioxiocom/firedantic/compare/0.2.8...0.3.0